requirements.txt
requests
pydantic
python-slugify
pybtex
rapidfuzz
pandas
bibtexparser
pyyaml
jinja2
lxml
